+++ internal/materia/host_manager.go
@@ -3,17 +3,24 @@ package materia
 import (
 	"context"
 
-	"primamateria.systems/materia/internal/containers"
 	"primamateria.systems/materia/pkg/components"
-	"primamateria.systems/materia/pkg/serviceman"
+	"primamateria.systems/materia/pkg/containers"
+	"primamateria.systems/materia/pkg/services"
 )
 
+type ServiceManager interface {
+	ApplyService(context.Context, string, services.ServiceAction, int) error
+	GetService(context.Context, string) (*services.Service, error)
+	RunOneshotCommand(context.Context, int, string, []string) error
+	WaitUntilState(context.Context, string, string, int) error
+}
+
 type HostManager interface {
-	serviceman.ServiceManager
 	containers.ContainerManager
 	components.ComponentReader
 	components.ComponentWriter
 	FactsProvider
+	ServiceManager
 	ListInstalledComponents() ([]string, error)
 	InstallScript(context.Context, string, []byte) error
 	RemoveScript(context.Context, string) error