+++ internal/materia/materia_config.go
@@ -11,9 +11,9 @@ import (
 	"primamateria.systems/materia/internal/attributes/age"
 	fileattrs "primamateria.systems/materia/internal/attributes/file"
 	"primamateria.systems/materia/internal/attributes/sops"
-	"primamateria.systems/materia/internal/services"
 	"primamateria.systems/materia/pkg/executor"
 	"primamateria.systems/materia/pkg/planner"
+	"primamateria.systems/materia/pkg/services"
 )
 
 var (