+++ pkg/executor/quadlet.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/charmbracelet/log"
 	"primamateria.systems/materia/internal/actions"
-	"primamateria.systems/materia/internal/containers"
 	"primamateria.systems/materia/pkg/components"
+	"primamateria.systems/materia/pkg/containers"
 )
 
 func cleanupNetwork(ctx context.Context, e *Executor, v actions.Action) error {